initgfx was changed as follows:
Calculate a config ID which is a MD5 checksum of the output of `sysctl
dev.vgapci | sort'. If the file /var/initgfx_config.id exists, check if
its content matches the config ID. If that's the case, load the modules
set via the rc.conf variable initgfx_kmods, and exit. If the file
/var/initgfx_config.id does not exist, or its content doesn't match the
config ID, try to auto-detect the accelerated graphics driver. Use xinit
to test if the X server starts, change initgfx_kmods, write a new xorg
config file, and save the new config ID if the test was successful.
Otherwise, as a fallback, write a xorg config file for the VESA or SCFB
driver depending on whether the system was booted via BIOS or EFI.

The code of the files intel_setup, nvidia_setup, and ati_amd_setup
are now part of initgfx. The device IDs are now included from
/etc/initgfx_device.db, and the xorg config templates are now included
from /etc/initgfx_xorg.cfg.