Re-work load_acpi to use kenv smbios
Before load_acpi used devinfo to determine on what kind of hardware it is running and it relied on strings that can be different for every new hardware generation.
kenv smbios should be more reliable and better to maintain.
While here, fix detection of some Thinkpads in the function to suppres loading of acpi_video.