Fix for hard link creation errors on unionfs.
Since the commit [1] fusefs checks if the inodes of the source and the target
file are identical when creating a hard link. If they differ, an error is
returned. This, however, is a problem when using unionfs. Since hard links
across different file systems are not possible, unionfs copies the source file
from the r/o directory to the writeable directory first, and then creates the
hard link there.

* Install a fusefs2 kernel module which is identical to fusefs except for the
inode check.
* Let /etc/rc.d/mount_uzip unload fusefs if loaded, then load fusefs2.

1. https://cgit.freebsd.org/src/commit/sys/fs/fuse/fuse_vnops.c?h=stable/14&id=0bef4927ea858bb18b6f679bc0a36cff264dc842