build: Allow alternative installed packages as BUILD_DEPENDENCIES
textproc/markdown conflicts with textproc/discount which is a dependency
of KDE Plasma 6. Trying to uninstall discount would also deinstall some
parts of plasma.
To get around this, add a program name to each package origin in
BUILD_DEPENDENCIES and check first if the program is found. If not,
install the package.
This also allows to have devel/git-lite installed instead of devel/git.